+++ game/src/Player.h
@@ -15,7 +15,7 @@ public:
     Player(glm::vec3 initialPosition, glm::vec3 initialDirection);
     void Update(GLFWwindow *window, int width, int height);
     glm::mat4 GetViewProjection() const;
-    glm::vec3 GetCurrentChunkPosition() const;
+    glm::ivec3 GetCurrentChunkPosition() const;
     glm::vec3 GetViewDirection() const;
     glm::vec3 m_Position;
     glm::vec2 m_Direction;
